Fix styles and move title logic into app.jsx
