+++ src/shared/Header.jsx
@@ -2,28 +2,11 @@ import { useState, useEffect } from 'react';
 import { NavLink, useLocation } from 'react-router';
 import styles from '../css-modules/Header.module.css';
 
-function Header() {
-  const location = useLocation();
-  const [pageTitle, setPageTitle] = useState('Todo List');
-
-  // Set dynamic title based on route
-  useEffect(() => {
-    switch (location.pathname) {
-      case '/':
-        setPageTitle('Todo List');
-        break;
-      case '/about':
-        setPageTitle('About');
-        break;
-      default:
-        setPageTitle('Not Found');
-        break;
-    }
-  }, [location]);
+function Header({title}) {
 
   return (
     <header className={styles.header}>
-      <h1 className={styles.heading}>{pageTitle}</h1>
+      <h1 className={styles.heading}>{title}</h1>
       <nav className={styles.nav}>
         <NavLink
           to="/"
